requirements.txt
flask==3.0.0
pdfplumber==0.11.9
openai>=1.12.0
python-dotenv==1.0.0
